initial refactor compiling
